2328: Use more svgs
