+++ build-configs/malte/index.ts
@@ -93,9 +93,8 @@ const webMalteBuildConfig: WebBuildConfigType = {
   appDescription: 'Guide of the Malteser Werke for Refugees. Digital. Multilingual. Free.',
   mainImprint,
   icons: {
-    appLogo: '/app-logo.png',
+    appLogo: '/app-logo.svg',
     appLogoMobile: '/app-icon-round.svg',
-    locationMarker: '/location-marker.svg',
     appleTouchIcon: '/apple-touch-icon.png',
     socialMediaPreview: '/social-media-preview.png',
     favicons: '/favicons/',
