+++ native/src/assets/index.ts
@@ -15,6 +15,7 @@ import ExternalLinkIcon from '../../../assets/icons/external-link.svg'
 import HappySmileyIcon from '../../../assets/icons/happy-smiley.svg'
 import LanguageIcon from '../../../assets/icons/language.svg'
 import LocationFixedIcon from '../../../assets/icons/location-fixed.svg'
+import LocationMarkerIcon from '../../../assets/icons/location-marker.svg'
 import LocationNotFixedIcon from '../../../assets/icons/location-not-fixed.svg'
 import LocationOffIcon from '../../../assets/icons/location-off.svg'
 import MailIcon from '../../../assets/icons/mail.svg'
@@ -53,6 +54,7 @@ export {
   EventThumbnailPlaceholder3,
   ExternalLinkIcon,
   LocationFixedIcon,
+  LocationMarkerIcon,
   LocationNotFixedIcon,
   LocationOffIcon,
   HappySmileyIcon,
