+++ web/src/index.ejs
@@ -145,6 +145,7 @@
 
         #splash img {
             height: 50%;
+            width: 50%;
             max-width: 100%;
             object-fit: contain;
         }
